requirement.txt
python=3.8
pytorch=1.12.0
torchvision=0.12
cuda=11.5
